core/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
neo4j==5.12.0
qdrant-client==1.5.4
langchain==0.0.300
celery==5.3.4
redis==5.0.0
GitPython==3.1.35
requests==2.31.0
pytest==7.4.2
httpx==0.25.0
